+++ core/src/main/java/info/nightscout/androidaps/plugins/constraints/versionChecker/AllowedVersions.kt
@@ -4,42 +4,8 @@ import org.joda.time.LocalDate
 import org.json.JSONArray
 import org.json.JSONException
 import org.json.JSONObject
-import java.lang.Exception
 
 class AllowedVersions {
-
-    fun generateSupportedVersions(): String =
-        JSONArray()
-            // Android API versions
-            .put(JSONObject().apply {
-                put("minAndroid", 1) // 1.0
-                put("maxAndroid", 23) // 6.0.1
-            })
-            .put(JSONObject().apply {
-                put("minAndroid", 24) // 7.0
-                put("maxAndroid", 25) // 7.1.2
-                put("supported", "2.6.2")
-            })
-            .put(JSONObject().apply {
-                put("minAndroid", 26) // 8.0
-                put("maxAndroid", 27) // 8.1
-                put("supported", "2.8.2")
-            })
-            .put(JSONObject().apply {
-                put("minAndroid", 28) // 9.0
-                put("maxAndroid", 99)
-                put("supported", "2.8.2")
-            })
-            // Version time limitation
-            .put(JSONObject().apply {
-                put("endDate", "2021-11-07")
-                put("version", "2.9.0-beta1")
-            })
-            .put(JSONObject().apply {
-                put("endDate", "2021-11-07")
-                put("version", "3.0-beta1")
-            })
-            .toString()
 
     fun findByApi(definition: String?, api: Int): JSONObject? {
         if (definition == null) return null
